% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart.R
\name{plotmz}
\alias{plotmz}
\title{plot dart ms data}
\usage{
plotmz(data, inscf = 5, ...)
}
\arguments{
\item{data}{matrix with row mz, column rt and cell intensity}

\item{inscf}{intensity cutoff in Log scale}
}
\description{
plot dart ms data
}
